Landing strip updates
Landing strip for all models now has a waypoint at the same coordinate
as the loiter waypoint. Added because the condition for moving onto the
next waypoint for LOITER_TO_ALT WP is only altitude based rather than
altitude and coordinate based. Wp will be placed at altitude of the
mission. If mission is under 90m or unplanned the waypoint will be
placed at 90m.